+++ nens_auth_client/conf.py
@@ -21,8 +21,6 @@ class NensAuthClientAppConf(AppConf):
         required = (
             "CLIENT_ID",  # Provided by AWS Cognito
             "CLIENT_SECRET",  # Provided by AWS Cognito
-            "REDIRECT_URI",  # Depends on the urls.py of your django site
-            "LOGOUT_REDIRECT_URI",  # Depends on the urls.py of your django site
             "LOGOUT_URL",  # N&S Global (full URL ending with /logout)
             "ISSUER",  # N&S Global (authorization server URL)
         )